[2.x] addressing review
