+++ server-test/src/test/scala/testpkg/ClientSubscriptionTest.scala
@@ -6,7 +6,7 @@
  * Licensed under Apache License 2.0 (see LICENSE)
  */
 
-package sbt
+package testpkg
 
 import scala.concurrent.duration.*
 import sbt.internal.langserver.SbtExecParams
