+++ server-test/src/test/scala/testpkg/QueuedNotificationTest.scala
@@ -5,7 +5,7 @@
  * Licensed under Apache License 2.0 (see LICENSE)
  */
 
-package sbt
+package testpkg
 
 import scala.concurrent.duration.*
 import sbt.internal.langserver.SbtExecParams
