+++ server-test/src/test/scala/testpkg/ServerCompletionsTest.scala
@@ -5,7 +5,7 @@
  * Licensed under Apache License 2.0 (see LICENSE)
  */
 
-package sbt
+package testpkg
 
 import sbt.protocol.{ CompletionParams, CompletionResponse }
 import sbt.protocol.codec.JsonProtocol.given
